+++ NEWS.md
@@ -1,5 +1,11 @@
 # googlesheets4 (development version)
 
+* `gs4_auth(subject =)` is a new argument that can be used with
+  `gs4_auth(path =)`, i.e. when using a service account. The `path` and
+  `subject` arguments are ultimately processed by
+  `gargle::credentials_service_account()` and support the use of a service
+  account to impersonate a normal user.
+
 * `gs4_scopes()` is a new function to access scopes relevant to the Sheets and
   Drive APIs. When called without arguments, `gs4_scopes()` returns a named
   vector of scopes, where the names are the associated short aliases.